Chapter10/ch10/ms/algo/mergesort.py

def sort(v):
    if len(v) <= 1:
        return v
    mid = len(v) // 2
    v1, v2 = sort(v[:mid]), sort(v[mid:])
    return merge(v1, v2)


def merge(v1, v2):
    """Merge two vectors together.

    Assumes v1 and v2 are sorted (ascending).  Resulting vector
    will also be sorted ascending.
    """
    v = []
    h = k = 0
    len_v1, len_v2 = len(v1), len(v2)
    while h < len_v1 or k < len_v2:
        if k == len_v2 or (h < len_v1 and v1[h] < v2[k]):
            v.append(v1[h])
            h += 1
        else:
            v.append(v2[k])
            k += 1
    return v
